Make miscellaneous updates to other apps, mostly for consistency
