+++ apps/admin/backend/src/server.ts
@@ -68,20 +68,28 @@ function constructDippedSmartCardAuthMachineState(
 
 function buildApi(auth: DippedSmartCardAuthApi, workspace: Workspace) {
   return grout.createApi({
-    getAuthStatus: () =>
-      auth.getAuthStatus(constructDippedSmartCardAuthMachineState(workspace)),
+    getAuthStatus() {
+      return auth.getAuthStatus(
+        constructDippedSmartCardAuthMachineState(workspace)
+      );
+    },
 
-    checkPin: (input: { pin: string }) =>
-      auth.checkPin(constructDippedSmartCardAuthMachineState(workspace), input),
+    checkPin(input: { pin: string }) {
+      return auth.checkPin(
+        constructDippedSmartCardAuthMachineState(workspace),
+        input
+      );
+    },
 
-    logOut: () =>
-      auth.logOut(constructDippedSmartCardAuthMachineState(workspace)),
+    logOut() {
+      return auth.logOut(constructDippedSmartCardAuthMachineState(workspace));
+    },
 
-    programCard: ({
+    programCard({
       userRole,
     }: {
       userRole: 'system_administrator' | 'election_manager' | 'poll_worker';
-    }) => {
+    }) {
       const electionDefinition =
         getMostRecentlyCreateElectionDefinition(workspace);
       assert(electionDefinition !== undefined);
@@ -96,8 +104,11 @@ function buildApi(auth: DippedSmartCardAuthApi, workspace: Workspace) {
       return auth.programCard({ electionHash }, { userRole });
     },
 
-    unprogramCard: () =>
-      auth.unprogramCard(constructDippedSmartCardAuthMachineState(workspace)),
+    unprogramCard() {
+      return auth.unprogramCard(
+        constructDippedSmartCardAuthMachineState(workspace)
+      );
+    },
   });
 }
