+++ apps/admin/frontend/src/app.test.tsx
@@ -237,8 +237,8 @@ test('authentication works', async () => {
       electionHash: eitherNeitherElectionDefinition.electionHash,
     }),
   });
-  mockApiClient.checkPin.expectCallWith({ pin: '111111' }).resolves();
   await screen.findByText('Enter the card security code to unlock.');
+  mockApiClient.checkPin.expectCallWith({ pin: '111111' }).resolves();
   fireEvent.click(screen.getByText('1'));
   fireEvent.click(screen.getByText('1'));
   fireEvent.click(screen.getByText('1'));
@@ -277,8 +277,8 @@ test('authentication works', async () => {
       electionHash: eitherNeitherElectionDefinition.electionHash,
     }),
   });
-  mockApiClient.checkPin.expectCallWith({ pin: '123456' }).resolves();
   await screen.findByText('Enter the card security code to unlock.');
+  mockApiClient.checkPin.expectCallWith({ pin: '123456' }).resolves();
   fireEvent.click(screen.getByText('1'));
   fireEvent.click(screen.getByText('2'));
   fireEvent.click(screen.getByText('3'));
