+++ apps/scan/backend/src/app_config.test.ts
@@ -8,22 +8,13 @@ import waitForExpect from 'wait-for-expect';
 import { LogEventId } from '@votingworks/logging';
 import * as grout from '@votingworks/grout';
 import {
-  ALL_PRECINCTS_SELECTION,
-  ReportSourceMachineType,
-  ScannerReportData,
-  ScannerReportDataSchema,
   SCANNER_RESULTS_FOLDER,
   singlePrecinctSelectionFor,
 } from '@votingworks/utils';
-import { assert, err, ok, Result } from '@votingworks/basics';
+import { assert, err, ok } from '@votingworks/basics';
 import fs from 'fs';
 import { join } from 'path';
-import {
-  fakeElectionManagerUser,
-  fakePollWorkerUser,
-  generateCvr,
-  mockOf,
-} from '@votingworks/test-utils';
+import { generateCvr } from '@votingworks/test-utils';
 import {
   ballotImages,
   configureApp,
@@ -353,60 +344,5 @@ test('auth', async () => {
     1,
     { electionHash },
     { pin: '123456' }
-  );
-});
-
-test('write scanner report data to card', async () => {
-  const { apiClient, mockAuth, mockUsb } = await createApp();
-  await configureApp(apiClient, mockUsb);
-
-  mockOf(mockAuth.writeCardData).mockImplementation(() =>
-    Promise.resolve(ok())
-  );
-
-  const { electionDefinition } = electionFamousNames2021Fixtures;
-  const { electionHash } = electionDefinition;
-  const scannerReportData: ScannerReportData = {
-    ballotCounts: {},
-    isLiveMode: false,
-    machineId: '0000',
-    pollsTransition: 'close_polls',
-    precinctSelection: ALL_PRECINCTS_SELECTION,
-    tally: [],
-    tallyMachineType: ReportSourceMachineType.PRECINCT_SCANNER,
-    timePollsTransitioned: 0,
-    timeSaved: 0,
-    totalBallotsScanned: 0,
-  };
-  let result: Result<void, Error>;
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({ status: 'logged_out', reason: 'no_card' })
-  );
-  result = await apiClient.saveScannerReportDataToCard({ scannerReportData });
-  expect(result).toEqual(err(new Error('User is not logged in')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakeElectionManagerUser(electionDefinition),
-    })
-  );
-  result = await apiClient.saveScannerReportDataToCard({ scannerReportData });
-  expect(result).toEqual(err(new Error('User is not a poll worker')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakePollWorkerUser(electionDefinition),
-    })
-  );
-  result = await apiClient.saveScannerReportDataToCard({ scannerReportData });
-  expect(result).toEqual(ok());
-  expect(mockAuth.writeCardData).toHaveBeenCalledTimes(1);
-  expect(mockAuth.writeCardData).toHaveBeenNthCalledWith(
-    1,
-    { electionHash },
-    { data: scannerReportData, schema: ScannerReportDataSchema }
   );
 });
